:recycle: Glog Color, PageUI Extension
