+++ Paper/Utils/Extensions/GlogColor.swift
@@ -1,0 +1,16 @@
+//
+//  GlogColor.swift
+//  Paper
+//
+//  Created by 선민재 on 2022/11/07.
+//
+
+import UIKit
+
+extension UIColor {
+    static var backgroundColor = UIColor(named: "backgroundColor")
+    static var Gradient1 = UIColor(named: "Gradient1")
+    static var Gradient2 = UIColor(named: "Gradient2")
+    static var textField = UIColor(named: "textField")
+    static var icon = UIColor(named: "icon")
+}
